refactor(memory): address PR review comments
- Remove create_storage_bundle, consolidate into create_storage returning
  StorageBundle directly (removes redundant tuple wrapper)
- Move MemoryRuntimeConfig from memory::context to config::types so
  config is self-contained without gateway dependencies
- Add doc comments to all MemoryExecutionContext fields explaining
  requested vs active semantics and each field's purpose
- Move memory header constants from memory/headers.rs to
  routers/common/header_utils.rs where other x-smg-* headers live
- Remove duplicated serde deserialization test (already covered in
  config::types::tests)
