+++ model_gateway/src/memory/mod.rs
@@ -1,7 +1,7 @@
 pub mod context;
 pub mod headers;
 
-pub use context::{MemoryExecutionContext, MemoryRuntimeConfig};
+pub use context::MemoryExecutionContext;
 pub use headers::{
     MemoryHeaderView, MEMORY_EMBEDDING_MODEL_HEADER, MEMORY_EXTRACTION_MODEL_HEADER,
     MEMORY_LTM_STORE_ENABLED_HEADER, MEMORY_POLICY_HEADER, MEMORY_RECALL_METHOD_HEADER,
