+++ model_gateway/src/routers/common/header_utils.rs
@@ -9,6 +9,15 @@ static HEADER_TARGET_WORKER: HeaderName = HeaderName::from_static("x-smg-target-
 static HEADER_ROUTING_KEY: HeaderName = HeaderName::from_static("x-smg-routing-key");
 static HEADER_MCP: HeaderName = HeaderName::from_static("x-smg-mcp");
 
+// Placeholder header names for memory controls.
+// These names are under active discussion and may change in a follow-up PR.
+pub const MEMORY_POLICY_HEADER: &str = "x-smg-memory-policy";
+pub const MEMORY_LTM_STORE_ENABLED_HEADER: &str = "x-smg-memory-ltm-store-enabled";
+pub const MEMORY_SUBJECT_ID_HEADER: &str = "x-smg-memory-subject-id";
+pub const MEMORY_RECALL_METHOD_HEADER: &str = "x-smg-memory-recall-method";
+pub const MEMORY_EMBEDDING_MODEL_HEADER: &str = "x-smg-memory-embedding-model";
+pub const MEMORY_EXTRACTION_MODEL_HEADER: &str = "x-smg-memory-extraction-model";
+
 fn extract_header_value<'a>(headers: Option<&'a HeaderMap>, name: &HeaderName) -> Option<&'a str> {
     headers
         .and_then(|h| h.get(name))
